changements divers

+++ srcs/mem/ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: snicolet <marvin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 12:18:16 by snicolet          #+#    #+#             */
-/*   Updated: 2015/11/24 13:34:32 by snicolet         ###   ########.fr       */
+/*   Updated: 2015/11/27 01:17:52 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-	unsigned int			p;
+	size_t					p;
 	unsigned const char		*s;
 	unsigned char			*d;
